+++ src/main/java/cart/controller/JwpCartController.java
@@ -1,17 +1,12 @@
 package cart.controller;
 
 import cart.dto.ProductDto;
-import cart.dto.ProductRequestDto;
 import cart.dto.ProductResponseDto;
 import cart.service.JwpCartService;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
 
-import javax.validation.Valid;
-import java.net.URI;
 import java.util.List;
 
 import static java.util.stream.Collectors.toList;
@@ -43,30 +38,5 @@ public class JwpCartController {
                 .collect(toList());
         model.addAttribute("products", response);
         return "admin";
-    }
-
-    @PostMapping("/admin/products")
-    @ResponseStatus(HttpStatus.CREATED)
-    public ResponseEntity<ProductResponseDto> addProduct(@RequestBody @Valid ProductRequestDto productRequestDto) {
-        ProductDto productdto = jwpCartService.add(productRequestDto);
-        ProductResponseDto response = ProductResponseDto.fromProductDto(productdto);
-        return ResponseEntity.status(HttpStatus.CREATED)
-                .location(URI.create("/admin/products/" + response.getId()))
-                .body(response);
-    }
-
-    @PutMapping("/admin/products/{id}")
-    public ResponseEntity<ProductResponseDto> updateProduct(@PathVariable("id") Long id,
-                                                            @RequestBody @Valid ProductRequestDto productRequestDto) {
-        ProductDto productDto = jwpCartService.updateById(productRequestDto, id);
-        ProductResponseDto response = ProductResponseDto.fromProductDto(productDto);
-        return ResponseEntity.ok().body(response);
-    }
-
-    @DeleteMapping("/admin/products/{id}")
-    public ResponseEntity<Void> deleteProduct(@PathVariable("id") Long id) {
-        jwpCartService.deleteById(id);
-        return ResponseEntity.ok()
-                .build();
     }
 }
