+++ src/main/java/cart/controller/JwpCartControllerAdvice.java
@@ -1,14 +1,12 @@
 package cart.controller;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-@ControllerAdvice
+@RestControllerAdvice
 public class JwpCartControllerAdvice {
 
-    @ResponseBody
     @ExceptionHandler(Exception.class)
     public ResponseEntity<String> handle(Exception e) {
         return ResponseEntity.badRequest().body(e.getMessage());
