+++ src/main/java/cart/repository/ProductRepository.java
@@ -1,5 +1,6 @@
-package cart.domain;
+package cart.repository;
 
+import cart.domain.Product;
 import cart.entity.ProductEntity;
 
 import java.util.List;
